Updated Page Layout
The basic page layout is a fixed bootstrap navbar.
The main area is using a grid layout with two cloumns: sidebar | content.

Sidebar:
   A <ul> that has the different options a user can click on to display the different areas of the app
   on hover each <li> will change font color and have a 50% opacity background
Content:
   This will be where the Content of the clicked <li> will be displayed.

The JS to handle the API calls for signing in/up/out/change-pw have also been tested and added, though they are not displayed anywhere in this commit.