Several changes come along with this commit.
I have finished the delete function 100%

Using a mixture of handlebars and jquery I am able to show only the
devices that belong to the current_user.

edit/update not fully functional. I was able to prepopulate the form fields wth
the data from the origional device, however when the api call is attempted it
only causes the pages to refresh.

minor styling has been applyed to the Handlebars templates.