Fixed Edit Device.
Tracked dow the issue where submitting the Update form would cause a page refresh.
The iss was cause by attempting to add an eventhandler on an element that is not present
when the handler is assigned.
By moving the eventhandler to the onShowEditDevice function I am now able to succefuly
update device info via the api.